refactor(runtime/llm): extract failed-tool-tracking helpers to own module
Moves the ~150 LOC "model keeps retrying the same broken call"
detection + recovery-hint machinery out of chat-executor-tool-loop.ts
into a new chat-executor-failed-tool-tracking.ts. The main loop file
shrinks from 2,761 -> 2,609 LOC; behavior is unchanged (all helpers
keep the same names + signatures + semantics).

Moved exports:
- FAILED_TOOL_RECOVERY_STREAK constant (= 3)
- summarizeToolFailureForRecovery
- buildFailedToolRecoveryHint
- toolFailureSignature
- toolCallSignature
- isRepeatedSameFailedToolPattern
- responseRepeatsFailedToolPattern
- collectRecentConsecutiveFailedToolCalls
- updateFailedToolStreak
- mergeRecoveryHints

stableToolFailureValue is kept file-private (only
toolFailureSignature needed it).

Purely a file-organization change — zero behavior delta. Typecheck
green. Full llm/ test suite (720 tests, 1 pre-existing failure on
main unchanged) still matches baseline.
